requirements.txt
numpy>=2.0.0
opencv-python>=4.10.0.84
onnxruntime-directml>=1.20.0
onnx>=1.16.0
scikit-learn>=1.5.0
skl2onnx>=1.17.0
Pillow>=10.4.0
